+++ _3DRepo_Toolkit_Adapter/3DRepo_Adapter.csproj
@@ -103,6 +103,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="AdapterActions\Push.cs" />
     <Compile Include="CRUD\Create\IObjects.cs" />
     <Compile Include="CRUD\Create\_Create.cs" />
     <Compile Include="CRUD\Delete\_Delete.cs" />
@@ -114,7 +115,7 @@
     <Compile Include="CRUD\Read\_Read.cs" />
     <Compile Include="Types\PanelPlanar.cs" />
     <Compile Include="CRUD\Update\_UpdateObjects.cs" />
-    <Compile Include="_3DRepo_ToolkitAdapter.cs" />
+    <Compile Include="3DRepo_Adapter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Types\Comparer.cs" />
     <Compile Include="Types\DependencyTypes.cs" />
@@ -130,9 +131,7 @@
       <Private>False</Private>
     </ProjectReference>
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="AdapterActions\" />
-  </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
